pyproject.toml
[project]
name = "git-assume"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Josuke Yamane", email = "josuke_yamane@mail.toyota.co.jp" }
]
dependencies = [
    "pytest>=8.3.3",
    "black>=24.8.0",
    "isort>=5.13.2",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
git-assume = "git_assume.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/git_assume"]
